cmd/contour: integrate contourcli into contour docker image
Updates #293

This PR moves `cmd/contourcli` into the main `cmd/contour` binary so
that it can be used via kubectl exec.
