+++ docs/troubleshooting.md
@@ -27,6 +27,22 @@ kubectl -n heptio-contour port-forward $CONTOUR_POD 9001
 ```
 Then navigate to [http://127.0.0.1:9001/](http://127.0.0.1:9001/) to access the admin interface for the Envoy container running on that pod.
 
+## Interrogate Contour's gRPC API
+
+Sometimes it's helpful to be able to interrogate Contour to find out exactly the data it is sending to Envoy.
+Contour ships with a `contour cli` subcommand which can be used for this purpose.
+
+Because Contour runs as a sidecar with Envoy, you must run `contour cli` commands _inside_ the pod.
+The easiest way do this is via `kubectl exec`
+```
+# Get one of the pods that matches the deployment/daemonset
+CONTOUR_POD=$(kubectl -n heptio-contour get pod -l app=contour -o jsonpath='{.items[0].metadata.name}')
+# Do the port forward to that pod
+kubectl -n heptio-contour exec $CONTOUR_POD -c contour contour cli lds
+```
+Which will stream changes to the LDS api endpoint to your terminal.
+Replace `contour cli lds` with `contour cli rds` for RDS, `contour cli cds` for CDS, and `contour cli eds` for EDS.
+
 ## Can't make kube-lego work with Contour
 
 If you use [kube-lego][0] for Let's Encrypt SSL certificates, kube-lego appears to set the ingress class on the ingress record it uses for the acme-01 challenge to `nginx`.
